aggiornati tipi dto

+++ src/main/java/com/angeloni/nutricare/dto/AnthropometryDto.java
@@ -11,7 +11,7 @@ public class AnthropometryDto implements Serializable {
 	private static final long serialVersionUID = 4183994057140521879L;
 	
 	@PositiveOrZero(message = "Value must be higher or equal to zero")
-	private Integer height;
+	private Double height;
 	@PositiveOrZero(message = "Value must be higher or equal to zero")
 	private Double weight;
 	private FoldDto fold;
